added upadte delete and read functionality

+++ src/app/api/(books-query)/delete-book/[bookId]/route.ts
@@ -3,15 +3,25 @@ import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
 
-export async function DELETE({ params }: { params: { bookId: string } }) {
+export async function DELETE(
+  req: Request,
+  { params }: { params: { bookId: string } }
+) {
   const session = await getServerSession(authOptions);
 
-  if (!session || !session.user?.id) {
+  if (!session) {
     return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
   }
-  const bookId = params.bookId;
+
+  const userId = session.user.id;
+  const { bookId } = await params;
+
   try {
-    await prisma.book.delete({ where: { id: bookId } });
+    const validUser = await prisma.user.findUnique({ where: { id: userId } });
+    if (!validUser) {
+      return NextResponse.json({ error: "No user found" }, { status: 404 });
+    }
+    await prisma.book.delete({ where: { id: bookId, userId: userId } });
     return NextResponse.json(
       { message: "Deleted successfully", success: true },
       { status: 201 }
